+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Allow for fun names starting with _ (for lambdas mostly)
 * Implement CEILING(n, m=1)
 * Implement FLOOR(n, m=1)
+* Implement ROUND(n, m=0)
 
 
 ## xel.js 1.3.0  released 2024-04-05